fifty_off/requirements.txt
arrow==0.15.4
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Django==3.0
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
Flask==1.1.1
geos==0.2.2
httpie==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
Pillow==6.2.1
proj==0.1.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
Werkzeug==0.16.0
